pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dj_drf_utils"]

[project]
name = "dj_drf_utils"
version = "1.3.4"
authors = [
    {name = "Cayo Rodrigues", email = "cayo.rodrigues1914@gmail.com"},
]
maintainers = [
    {name = "Cayo Rodrigues", email = "cayo.rodrigues1914@gmail.com"},
]
description = "Useful functions and classes for Django and Django Rest Framework"
readme = "README.md"
requires-python = ">=3.8"
keywords = [
    "django", "utils", "serializers", "generic", "views", "viewsets", "mixins",
    "email", "login", "model", "manager", "custom", "action", "shortcut", "error",
    "simple", "rest", "framework", "dj", "drf", "admin", "hash", "password", "user",
    "filter", "queryset",
]
license = {text = "MIT License"}
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "djangorestframework",
    "django-custom-cache-page",
]

[project.urls]
repository = "https://github.com/cayo-rodrigues/dj-drf-utils"